requirements.txt
python-telegram-bot==20.0 
qrcode==7.3 